+++ mex/common/types/text.py
@@ -20,6 +20,30 @@ class TextLanguage(StrEnum):
     RU = "ru"
 
 
+class RestrictedTextLanguage(StrEnum):
+    """Allowes only English and German as language tags for `Text` values."""
+
+    DE = "de"
+    EN = "en"
+
+
+def get_language_confidence(detector: DetectorFactory) -> RestrictedTextLanguage | None:
+    """Assigns None as Language if confidence is below 0.75 or is not En or DE."""
+    probs = detector.get_probabilities()
+    if not probs:
+        return None
+
+    best = max(probs, key=lambda p: p.prob)
+    confidence_threshold = 0.75
+
+    if best.prob < confidence_threshold:
+        return None
+    try:
+        return RestrictedTextLanguage(best.lang)
+    except ValueError:
+        return None
+
+
 class Text(BaseModel):
     """Type class for text objects.
 
@@ -30,7 +54,7 @@ class Text(BaseModel):
     """
 
     value: Annotated[str, Field(min_length=1)]
-    language: TextLanguage | None = None
+    language: RestrictedTextLanguage | TextLanguage | None = None
 
     @model_validator(mode="before")
     @classmethod
@@ -42,7 +66,7 @@ class Text(BaseModel):
             try:
                 detector = DETECTOR_FACTORY.create()
                 detector.append(value)
-                language = TextLanguage(detector.detect())
+                language = get_language_confidence(detector)
             except (LangDetectException, ValueError):
                 pass
         return {"language": language, "value": value}